Spanish Vocab - Using FastAPI - OpenAPI - ChatGPT3-5 LLM/requirements.txt
flask==3.0.2
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
